pm-hotplug: do not consider frequency in the cpu hotplug heuristic
Do not consider the cpu0 frequency in the heuristic to decide if we need
to turn on/off the secondary cpu (cpu1). This works only if we're using
a dynamic scaling governor (ondemand) and not for custom governors or
governors that have a fixed range of frequency (i.e, performance or
powersave).

Signed-off-by: Andrea Righi <andrea@betterlinux.com>